refactor: split file manager operation routes
